move old python based code to 'slow' submodule.
