+++ forcepho/slow/gaussmodel.py
