+++ forcepho/slow/stamp.py
